Add files via upload
I am excited to share an ambitious and innovative project that has been a highlight of our engineering journey:the design and construction of an RC aeroplane for SAEISS Design and Development Competition (DDC) competition.

We created sketches, 3D models, and preliminary layouts to visualize the design using AutoCAD and Onshape. We also used modelling tools like MATLAB and Ansys as a analysis software to maximize the proficiency and output of the plane . 

It has provided invaluable hands-on experience, bridging the gap between theoretical knowledge and practical application. The project has also showcased the potential of collaborative engineering efforts in advancing aerospace technology. Looking ahead, we aim to leverage the insights gained from this project to explore new avenues in aviation. Our focus will be on developing more advanced aircraft, incorporating cutting-edge technologies such as electric propulsion, autonomous flight systems, and smart materials.